Update: メモ削除処理追加

+++ app/controllers/users/people_controller.rb
@@ -33,9 +33,7 @@ class Users::PeopleController < ApplicationController
   def destroy
     person = Person.find(params[:id])
 
-    if person.user_id == current_user.id
-      person.destroy
-    end
+    person.destroy if person.user_id == current_user.id
 
     redirect_to users_people_path
   end
